requirements.txt
dj_database_url
bootstrap4
Django
djangorestframework
djangorestframework-jwt
django-extensions
django-bootstrap4
django-environ
gunicorn
lxml
psycopg2
